Void_Cloud_Intelligence/void_ai_earth/requirements.txt
accelerate==1.9.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
asyncio==3.4.3
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
bidict==0.23.1
blobfile==3.0.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
chromadb==1.0.15
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
datasets==4.0.0
Deprecated==1.2.18
deprecation==2.1.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
docutils==0.21.2
durationpy==0.10
fastapi==0.115.13
filelock==3.18.0
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.3.0
google-auth==2.40.3
google-genai==1.28.0
googleapis-common-protos==1.70.0
gotrue==2.12.3
grpcio==1.73.1
gunicorn==21.2.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.3
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
ifaddr==0.2.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
inquirerpy==0.3.4
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kubernetes==33.1.0
limits==5.4.0
llama_cpp_python==0.3.9
lxml==6.0.0
markdown-it-py==3.0.0
markdown2==2.5.3
MarkupSafe==3.0.2
mdurl==0.1.2
mmh3==5.1.0
mpmath==1.3.0
multidict==6.6.3
multiprocess==0.70.16
networkx==3.5
nicegui==2.21.1
numpy==2.3.1
oauthlib==3.3.1
onnxruntime==1.22.1
opentelemetry-api==1.35.0
opentelemetry-exporter-otlp-proto-common==1.35.0
opentelemetry-exporter-otlp-proto-grpc==1.35.0
opentelemetry-proto==1.35.0
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0
orjson==3.11.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==25.0
pandas==2.3.1
pefile==2023.2.7
peft==0.16.0
pfzy==0.3.4
pillow==11.3.0
postgrest==1.1.1
posthog==5.4.0
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
pscript==0.7.7
psutil==7.0.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.1
pycparser==2.22
pycryptodomex==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyinstaller==6.14.2
pyinstaller-hooks-contrib==2025.6
PyJWT==2.10.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.2
python-multipart==0.0.20
python-socketio==5.13.0
pytz==2025.2
pywin32-ctypes==0.2.3
PyYAML==6.0.2
realtime==2.6.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.0.0
rpds-py==0.26.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.16.0
selenium==4.34.2
sentence-transformers==5.0.0
setuptools==80.9.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.46.2
storage3==0.12.0
StrEnum==0.4.15
supabase==2.17.0
supafunc==0.10.1
sympy==1.14.0
tenacity==8.5.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.2
torch==2.7.1
tqdm==4.67.1
transformers==4.53.3
trio==0.30.0
trio-websocket==0.12.2
tweepy==4.16.0
typer==0.16.0
typing==3.7.4.3
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.34.3
vbuild==0.8.2
watchfiles==1.1.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.2
wsproto==1.2.0
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
